test(presentation): refactor tests of cardlist creating helper functions
