+++ tests/presentation/helpers/testFactories.ts
@@ -1,0 +1,3 @@
+import { faker } from '@faker-js/faker';
+
+export const makeTitle = () => faker.random.word();
